requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
click==8.1.3
Flask==2.1.2
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.4.0
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
tomli==2.0.1
Werkzeug==2.1.2
zipp==3.8.0
